Move connected NotebookNativePreview to query_builder
