+++ frontend/src/metabase/querying/notebook/components/NotebookNativePreview/NotebookNativePreview.tsx
@@ -5,11 +5,7 @@ import { useGetNativeDatasetQuery } from "metabase/api";
 import { DelayedLoadingSpinner } from "metabase/common/components/EntityPicker/components/LoadingSpinner";
 import { color } from "metabase/lib/colors";
 import { getEngineNativeType } from "metabase/lib/engine";
-import { useDispatch, useSelector } from "metabase/lib/redux";
-import { checkNotNull } from "metabase/lib/types";
-import { setUIControls, updateQuestion } from "metabase/query_builder/actions";
 import { CodeMirrorEditor as Editor } from "metabase/query_builder/components/NativeQueryEditor/CodeMirrorEditor";
-import { getQuestion } from "metabase/query_builder/selectors";
 import { Box, Button, Flex, Icon, rem } from "metabase/ui";
 import * as Lib from "metabase-lib";
 import type Question from "metabase-lib/v1/Question";
@@ -34,29 +30,7 @@ const BUTTON_TITLE = {
   },
 };
 
-export const NotebookNativePreview = () => {
-  const dispatch = useDispatch();
-  const question = checkNotNull(useSelector(getQuestion));
-
-  const handleConvertClick = useCallback(
-    (newQuestion: Question) => {
-      dispatch(
-        updateQuestion(newQuestion, { shouldUpdateUrl: true, run: true }),
-      );
-      dispatch(setUIControls({ isNativeEditorOpen: true }));
-    },
-    [dispatch],
-  );
-
-  return (
-    <ControlledNotebookNativePreview
-      question={question}
-      onConvertClick={handleConvertClick}
-    />
-  );
-};
-
-export const ControlledNotebookNativePreview = ({
+export const NotebookNativePreview = ({
   question,
   onConvertClick,
   buttonTitle,
